add UILabel

+++ engine/module/runtime/Runtime.cpp
@@ -166,16 +166,16 @@ Color luax_to_color(lua_State* L, int index)
     luaL_checktype(L, index, LUA_TTABLE);
     Color color;
     lua_getfield(L, index, "r");
-    color.r = (int)lua_tointeger(L, -1);
+    color.r = (uint8_t)lua_tointeger(L, -1);
     lua_pop(L, 1);
     lua_getfield(L, index, "g");
-    color.g = (int)lua_tointeger(L, -1);
+    color.g = (uint8_t)lua_tointeger(L, -1);
     lua_pop(L, 1);
     lua_getfield(L, index, "b");
-    color.b = (int)lua_tointeger(L, -1);
+    color.b = (uint8_t)lua_tointeger(L, -1);
     lua_pop(L, 1);
     lua_getfield(L, index, "a");
-    color.a = (int)lua_tointeger(L, -1);
+    color.a = (uint8_t)lua_tointeger(L, -1);
     lua_pop(L, 1);
     return color;
 }
